refactor: 불필요한 getCartSummary 함수 제거
