+++ src/advanced/components/cart-section/product-select/StockStatus.tsx
@@ -1,9 +1,9 @@
 import { useCartState } from '../../../contexts/CartContext';
-import { getCartSummary } from '../../../reducer';
+import { getStockMessages } from '../../../reducer';
 
 export const StockStatus = () => {
   const state = useCartState();
-  const { stockMessages } = getCartSummary(state);
+  const stockMessages = getStockMessages(state);
 
   return (
     <div id='stock-status' className='text-xs text-red-500 mt-3 whitespace-pre-line'>
