+++ src/advanced/components/order-summary/summary-details/ItemDetails.tsx
@@ -1,12 +1,15 @@
 import { useCartState } from '../../../contexts/CartContext';
-import { getCartSummary } from '../../../reducer';
-import { CartItemForDisplay } from '../../../types';
+import { getCartDetails } from '../../../reducer';
 
 export const ItemDetails = () => {
   const state = useCartState();
-  const { cartItemsForDisplay } = getCartSummary(state);
+  const cartItemsForDisplay = getCartDetails(state).map((item) => ({
+    name: item.product ? item.product.name : '',
+    quantity: item.quantity,
+    totalPrice: item.itemTotal,
+  }));
 
-  return cartItemsForDisplay.map((item: CartItemForDisplay) => (
+  return cartItemsForDisplay.map((item) => (
     <div key={item.name} className='flex justify-between text-xs tracking-wide text-gray-400'>
       <span>
         {item.name} x {item.quantity}
